Change package names to match the new ownership of the project
