+++ app/com/codelab27/cards9/game/GameTypes.scala
@@ -1,0 +1,10 @@
+package com.codelab27.cards9.game
+
+object GameTypes {
+  type PlayerId = Int
+  type MatchId = Int
+  type CardId = Int
+  type CardClassId = Int
+  type FightId = Int
+  type Score = (Int, Int)
+}
