+++ test/com/codelab27/cards9/specs/ModelSpec.scala
@@ -1,0 +1,6 @@
+package com.codelab27.cards9.specs
+
+import org.scalatest._
+import org.scalatest.prop.GeneratorDrivenPropertyChecks
+
+abstract class ModelSpec extends WordSpec with Matchers with ShouldMatchers with GeneratorDrivenPropertyChecks
