--- test/specs/ModelSpec.scala
@@ -1,6 +1,0 @@
-package specs
-
-import org.scalatest._
-import org.scalatest.prop.GeneratorDrivenPropertyChecks
-
-abstract class ModelSpec extends WordSpec with Matchers with ShouldMatchers with GeneratorDrivenPropertyChecks
